requirements.txt
Flask
Flask-RESTful
Flask-JWT
Flask-SQLAlchemy
Marshmallow==3.0.0rc6
uwsgi
mysql-connector
virtualenv
pytest
coverage
pytest-cov
flask-cors
pyjwt==1.4.0